extract validateFormSection function; update tests
